requirements.txt
blosc2==2.2.3
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tqdm==4.65.0
